BlogGeneration-API-web/pyproject.toml
[project]
name = "bloggeneration-api-copy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.3.0",
    "duckduckgo-search>=8.1.1",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.1",
    "ipython>=9.7.0",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-google-genai>=3.0.1",
    "langchain-groq>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.3",
    "langgraph-cli[inmem]>=0.4.7",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "torch>=2.9.0",
    "typing-extensions>=4.15.0",
    "uvicorn>=0.38.0",
    "watchdog>=6.0.0",
    "wikipedia>=1.4.0",
]
